requirements.txt
# Sysrai Platform Requirements
# Main platform dependencies (Digital Ocean)

# Core Web Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.12.1

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-decouple==3.8
pydantic[email]==2.5.2
pydantic-settings==2.1.0

# Payments
stripe==7.4.0

# Redis & Caching
redis==5.0.1
celery==5.3.4

# Cloud Storage
boto3==1.34.0

# RunPod API Client
httpx==0.25.2
aiohttp==3.9.1

# AI & Language Models
langchain-openai==0.1.8
openai==1.3.8

# EPUB Processing
ebooklib==0.18
beautifulsoup4==4.12.2

# Video Processing (for local testing)
moviepy==1.0.3
opencv-python-headless==4.8.1.78

# Utilities
python-dotenv==1.0.0
pendulum==3.0.0
requests==2.31.0

# Development & Testing
pytest==7.4.3
pytest-asyncio==0.21.1
playwright==1.40.0

# Monitoring
sentry-sdk[fastapi]==1.38.0